rename & doc fix
