+++ models/gen_image/__init__.py
@@ -1,5 +1,5 @@
 """Image-generation model wrappers."""
 
-from models.gen_image.qwen_edit import QwenEditModel
+from models.gen_image.qwen_edit_model import QwenEditModel
 
 __all__ = ["QwenEditModel"]
